Emit PHP 8 attributes as comments in PHP 7
