+++ src/main/php/lang/ast/emit/PHP70.class.php
@@ -13,6 +13,7 @@ class PHP70 extends PHP {
   use 
     ArbitrayNewExpressions,
     ArrayUnpackUsingMerge,
+    AttributesAsComments,
     MatchAsTernaries,
     NullsafeAsTernaries,
     OmitArgumentNames,
