+++ src/main/php/lang/ast/emit/PHP71.class.php
@@ -11,6 +11,7 @@ class PHP71 extends PHP {
   use 
     ArbitrayNewExpressions,
     ArrayUnpackUsingMerge,
+    AttributesAsComments,
     CallablesAsClosures,
     MatchAsTernaries,
     NonCapturingCatchVariables,
