+++ src/main/php/lang/ast/emit/PHP72.class.php
@@ -11,6 +11,7 @@ class PHP72 extends PHP {
   use 
     ArbitrayNewExpressions,
     ArrayUnpackUsingMerge,
+    AttributesAsComments,
     CallablesAsClosures,
     MatchAsTernaries,
     NonCapturingCatchVariables,
